Split converter app and library into separate source folders
