+++ source/CMakeLists.txt
@@ -5,5 +5,6 @@ add_subdirectory(lib_render)
 add_subdirectory(lib_image)
 add_subdirectory(lib_assetdb)
 add_subdirectory(lib_ecs)
+add_subdirectory(lib_recon)
 add_subdirectory(app_shell)
 add_subdirectory(app_recon)
