+++ source/lib_recon/private/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(potato_librecon PRIVATE
+    "${CMAKE_CURRENT_LIST_DIR}/context.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/converter_app.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/converter_config.cpp"
+)
+
+add_subdirectory(converters)
