+++ source/lib_recon/public/potato/recon/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(potato_librecon PRIVATE
     "${CMAKE_CURRENT_LIST_DIR}/context.h"
     "${CMAKE_CURRENT_LIST_DIR}/converter.h"
+    "${CMAKE_CURRENT_LIST_DIR}/converter_app.h"
     "${CMAKE_CURRENT_LIST_DIR}/converter_config.h"
 )
